requirements.txt
pandas==2.1.1
scikit-learn==1.3.1
tensorflow==2.14.0
pygame==2.5.2
pyyaml==6.0.1